refactoring, flags added

+++ HW12/temp_api.h
@@ -3,8 +3,8 @@
 #include "data.h"
 #include "list.h"
 
-void TEMP_API_StatisticsForTheMonth(list_t *list, int year_number,
-                                    int monthly_number);
+void TEMP_API_StatisticsForTheMonth(list_t *list, long int year_number,
+                                    long int monthly_number);
 void TEMP_API_StatisticsForTheYear(list_t *list, int year_number);
 
 #endif /*TEMP_API_H_*/
